requirements.txt
numpy~=2.2.0
opencv-python~=4.10.0.84
Flask~=3.1.0
pillow~=11.0.0
scikit-image~=0.25.0
setuptools~=75.6.0
scikit-learn~=1.6.0
requests~=2.32.3
mysql~=0.0.3
mysql-connector-python~=9.1.0